Extract preserved subcircuit trace SRJ logic
